Got a very basic JavaFX to work and Map the board to it
If you just run Output.java, it will print out the GUI with a checkered box which is the map space.
To show the map I've created, in Output.Java, in the method drawGame(), clear the double for loops and uncomment the block of text beginning with Tile[][] board = [...]
Then run the MainMapTester.java, inside the main function it should run create2dArray then Output.main. This is an unorthodox method of running a Javafx class and will return an error but it will work however a fix will be needed.
This is only to show what the GUI could look like